requirements.txt
asgiref
certifi
cffi
charset-normalizer
cryptography
defusedxml
dj-database-url
dj-rest-auth
Django
django-allauth
django-cors-headers
django-dotenv
djangorestframework
dotenv
drf-yasg
idna
inflection
oauthlib
packaging
psycopg2
pycparser
PyJWT
python-dotenv
python3-openid
pytz
PyYAML
requests
requests-oauthlib
sqlparse
typing_extensions
tzdata
uritemplate
urllib3
whitenoise
